Cleaned up compiler warnings

+++ RobotCSimulator/Source/ROBOTC/Program/Globals.cpp
@@ -13,7 +13,7 @@ tMotor rightMotorPort;
 tSensors rightEncoderPort;
 
 ubyte taskDelay;
-double controllerSpeed;
+float controllerSpeed;
 
 
 // Setters
@@ -33,7 +33,7 @@ void SetRightEncoder(tSensors port) {
 	rightEncoderPort = port;
 }
 
-void SetControllerSpeed(double speed) {
+void SetControllerSpeed(float speed) {
 	if(speed <= 1) {
 		controllerSpeed = speed;
 	}
