+++ RobotCSimulator/Source/ROBOTC/Program/Globals.h
@@ -50,7 +50,7 @@ void SetRightEncoder(tSensors port);
  *
  * Set the maximum speed of the robot. Only applies to motors that have been slewed.
 */
-void SetControllerSpeed(double speed);
+void SetControllerSpeed(float speed);
 
 
 /*
